requirements.txt
langchain
langchain_openai
langchain_community
langchain_experimental
langchain_google_genai
flashrank
openai
PyPDF2
requests
streamlit
fastapi
pydantic
python-dotenv
uvicorn
psycopg2
python-multipart
streamlit_float
unstructured
faiss-cpu
rank_bm25
python-docx